// types/personSocials.ts
export interface PersonSocials {
  id: number
  freebase_mid: string | null
  freebase_id: string | null
  imdb_id: string | null
  tvrage_id: number | null
  wikidata_id: string | null
  facebook_id: string | null
  instagram_id: string | null
  tiktok_id: string | null
  twitter_id: string | null
  youtube_id: any
}